Refactor teacher dashboard work to simplify `dataflow` merge [#169195585] - move CLUE-specific files into `src/clue` folder - eliminate references to CLUE-specific files from common code files
