+++ src/clue/components/teacher/sixpack-right-controls.sass
@@ -1,4 +1,4 @@
-@import ../../components/vars
+@import ../../../components/vars
 
 .sixpack-right-controls
   display: flex
