+++ src/clue/components/teacher/teacher-dashboard.sass
@@ -1,4 +1,4 @@
-@import ../vars
+@import ../../../components/vars
 
 $tab-height: 20px
