+++ src/clue/components/teacher/teacher-dashboard.tsx
@@ -1,6 +1,6 @@
 import { inject, observer } from "mobx-react";
 import * as React from "react";
-import { BaseComponent, IBaseProps } from "../base";
+import { BaseComponent, IBaseProps } from "../../../components/base";
 import { TeacherGroupTabComponent } from "./teacher-group-tab";
 import { TeacherStudentTabComponent } from "./teacher-student-tab";
 import "./teacher-dashboard.sass";
