+++ src/clue/components/teacher/teacher-group-six-pack.sass
@@ -1,4 +1,4 @@
-@import ../vars
+@import ../../../components/vars
 
 .teacher-group-six-pack
   display: grid
@@ -81,12 +81,12 @@
         width: 39px
         height: 30px
       &:hover
-          background-color: hsla(1, 0%, 0%, 0.1)
+        background-color: hsla(1, 0%, 0%, 0.1)
       .support
-        background-image: url("../../assets/icons/support/support.svg")
+        background-image: url("../../../assets/icons/support/support.svg")
         &:hover
-          background-image: url("../../assets/icons/support/support-hover.svg")
+          background-image: url("../../../assets/icons/support/support-hover.svg")
         &:active
-          background-image: url("../../assets/icons/support/support-active.svg")
+          background-image: url("../../../assets/icons/support/support-active.svg")
       .expand-group-view
-        background-image: url("../../assets/icons/clue-dashboard/expand-group-view.svg")
+        background-image: url("../../../assets/icons/clue-dashboard/expand-group-view.svg")
