+++ src/clue/components/teacher/teacher-group-tab.sass
@@ -1,4 +1,4 @@
-@import ../vars
+@import ../../../components/vars
 
 .teacher-group-tab
   display: grid
