+++ src/clue/components/teacher/teacher-student-tab.sass
@@ -1,4 +1,4 @@
-@import ../vars
+@import ../../../components/vars
 
 .teacher-student-tab
   .user-list
