+++ src/clue/components/teacher/teacher-support.sass
@@ -1,4 +1,4 @@
-@import ../vars
+@import ../../../components/vars
   
 .teacher-support
   display: flex
