+++ src/clue/components/teacher/teacher-supports.sass
@@ -1,4 +1,4 @@
-@import ../vars
+@import ../../../components/vars
   
 .teacher-supports
   .dash-header
